add compoenent folder
