--- FhirBlaze.PatientModule/Questionnaire.razor
@@ -1,7 +1,0 @@
-﻿@using FhirBlaze.SharedComponents
-@using Hl7.Fhir.Model
-
-<h3>Questionnaire</h3>
-
-<div>This is questionnaire</div>
-
